Move tests package containing classes being tested
Package private members can now be used from tests
