+++ vespaclient-container-plugin/src/test/java/com/yahoo/vespa/http/server/CollectingMetric.java
@@ -1,5 +1,5 @@
 // Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.feedhandler.v3;
+package com.yahoo.vespa.http.server;
 
 import com.yahoo.jdisc.Metric;
